refactor(condo): DOMA-9619 got rid of a cyclic dependency
